# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tascálogo API - Restaurant tracker for Portugal
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Restaurant {
  id: number;
  name: string;
  concelho: string;
  district: string;
  cuisine?: string;
  /**
   * @minimum 1
   * @maximum 5
   */
  rating?: number;
  notes?: string;
  visitDate?: string;
  createdAt?: string;
}

export interface CreateRestaurant {
  name: string;
  concelho: string;
  district: string;
  cuisine?: string;
  /**
   * @minimum 1
   * @maximum 5
   */
  rating?: number;
  notes?: string;
  visitDate?: string;
}

export interface WishlistItem {
  id: number;
  name: string;
  concelho: string;
  district: string;
  cuisine?: string;
  notes?: string;
  createdAt?: string;
}

export interface MarkVisitedBody {
  /**
   * @minimum 1
   * @maximum 5
   */
  rating?: number;
  notes?: string;
  visitDate?: string;
}

export interface CreateWishlistItem {
  name: string;
  concelho: string;
  district: string;
  cuisine?: string;
  notes?: string;
}

export type StatsTopCuisinesItem = {
  cuisine: string;
  count: number;
};

export type StatsMostVisitedConcelhosItem = {
  concelho: string;
  count: number;
};

export interface Stats {
  totalRestaurants: number;
  totalConcelhos: number;
  totalWishlist: number;
  avgRating: number;
  topCuisines: StatsTopCuisinesItem[];
  mostVisitedConcelhos: StatsMostVisitedConcelhosItem[];
}

export type ListRestaurantsParams = {
  /**
   * Filter by concelho
   */
  concelho?: string;
};

export type ListWishlistParams = {
  /**
   * Filter by concelho
   */
  concelho?: string;
};