# lib/api-zod/src/generated/types/markVisitedBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tascálogo API - Restaurant tracker for Portugal
 * OpenAPI spec version: 0.1.0
 */

export interface MarkVisitedBody {
  /**
   * @minimum 1
   * @maximum 5
   */
  rating?: number;
  notes?: string;
  visitDate?: string;
}